update and reuse
