+++ src/components/footer/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./footer";
+export { Footer } from "./footer";
