+++ src/layout/app/layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "@remix-run/react";
-import Footer from "components/footer";
+import { Footer } from "components/footer";
 import Header from "./header";
 
 export default function Layout() {
