+++ src/pages/admin/layout.tsx
@@ -2,7 +2,7 @@ import type { Endow } from "@better-giving/endowment";
 import { useLoaderData } from "@remix-run/react";
 import type { LoaderFunction } from "@vercel/remix";
 import { plusInt } from "api/schema/endow-id";
-import Footer from "components/footer";
+import { Footer } from "components/footer";
 import { appRoutes } from "constants/routes";
 import Layout from "layout/dashboard";
 import { CircleAlert } from "lucide-react";
